requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.1
idna==2.10
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
